Do not verify interactivity in tapWhenReady function
With Xcode 15 there is an issue where some elements report that they are not hittable but the tap method still works on them.
Since tapWhenReady is intended to be used to make sure that element will be correctly tapped on we do not need interactivity check here. Instead only wait for the element's existence.

To get same results as previous implementation of tapWhenReady call waitForInteractivity before it.